pyproject.toml
[project]
name = "joy-talker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hidapi>=0.15.0",
    "pynput>=1.8.1",
]
